Added inline-editing
