+++ package.json
@@ -12,6 +12,7 @@
     "react-alert": "^5.4.0",
     "react-alert-template-basic": "^1.0.0",
     "react-dom": "^16.8.4",
+    "react-inline-edition": "^1.4.2",
     "react-moment": "^0.8.4",
     "react-redux": "^6.0.1",
     "react-router-dom": "^5.0.0",
