--- src/components/Sidebar/index.js
@@ -1,2 +1,0 @@
-import Sidebar from './Sidebar';
-export default Sidebar;
